Move markerArrow to helpers.py
